Did some basic tower logic for testing
Aaron: The list of towers isn't actually a list of towers, it's a list of pointers. I think it'll be more helpful that way later. I'll explain how to use it properly later.

I've made detection for when an enemy is within range of a tower, now firing something at the enemy is next.